test: Fix mz-ore test flake
Some tests in mz-ore were setting and removing panic hooks. This had
the potential of interfering with concurrently running tests. This
commit moves these tests to their own files so that they aren't run
concurrently with any other test.

Fixes #16337
